Stricter encoding of inspected strings.
Rather than try to display the raw string value as-is, the inspector presents
string values as they would be written in JavaScript. This allows you to copy-
and-paste a string value from the inspector into JavaScript without needing to
escape special characters by hand.

To improve the readability of strings with newlines, in the special case that a
string is being inspected directly (rather than as a descendant of some object),
we check whether the string requires less escaping as a template literal, and if
so, present it as such.

Fixes observablehq/notebook#594.